requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml_html_clean
python-dotenv==1.0.1
yt-dlp>=2024.7
markdown-it-py==3.0.0
browser-cookie3==0.19.1
cryptography>=43.0.0
playwright>=1.43.0
